Fallback on default naming context when root naming context not available
When querying, the root naming context (top level domain) is used instead of the default one. In some cases, such as a child domain in a forest, we don't have access to the root domain, so trying to query the root naming context will result in raising an error: "Could not find domain DC=contenso,DC=com".

Here we still try to go for the root domain, but fallback on the default if it doesn't work. It may be nice to also add an option to specify the path we want to work with.