Allow weak TLS ciphers in LDAP connection
This is similar to the change in the Impacket repository:
https://github.com/SecureAuthCorp/impacket/pull/1449

Some domain controllers are using very weak ciphers, which are rejected
by the ssl library of Python3.10 by default. This sets the cipher
settings to the lowest value. Just like the certificate is currently not
validated at all, we don't really care about weak ciphers.

The same caveat about this bug in ldap3 applies:
https://github.com/cannatag/ldap3/pull/1067