parse: implement offline certificate template enumeration
This is based on The Registry Rundown research by Outflank's
Cedric Van Bockhaven and Max Grim as presented at Troopers 24.

Given the output of TrustedSec's reg_query BOF of the certificate
template cache as found in the registry, parse template properties
and analyze them for possible vulnerabilities.

 reg_query_recursive HKU .DEFAULT\Software\Microsoft\Cryptography\CertificateTemplateCache

The introduced parse command works completely offline and does not
connect to any remote system. Therefore, CA information is not available
and the reported data might be outdated.

As a result, it is unknown whether a given template is actually
enabled/published. Manually query the `certificateTemplates` attribute
of objects with class `pKIEnrollmentService` over LDAP and supply the
comma separated list to the `-published` command line option.

 ldapsearch "(objectclass=pKIEnrollmentService)" --attributes certificateTemplates --dn "CN=Configuration,DC=testlab,DC=local"

Similarly, the `-vulnerable` option requires a list of SIDs you control.
These can be specified as a comma separated list to the `-sids` option.

For testing purposes, the sorted JSON output of the regular `certipy find`
and newly introduced `certipy parse` can be compared using the following
jq(1) one-liner:

 jq '[ ."Certificate Templates"|values[] ] | sort_by(."Template Name")'