Removed Windows-specific functionality such as SSPI and PTT
The sheer size, third-party code, and recurring issues makes it hard to maintain for now. As of now, it's unknown how much this feature was used. Alternatives such as Rubeus is a perfect replacement for this functionality, and Rubeus is better maintained than the SSPI and PTT parts of this codebase.